% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_from_matrix}
\alias{convert_from_matrix}
\title{Convert a wide matrix to a tidy tibble}
\usage{
convert_from_matrix(data, samples_in_cols = TRUE)
}
\arguments{
\item{data}{A feature table matrix in wide format. To convert a wide data frame, see \code{\link[metamorphr]{convert_from_wide}}.}

\item{samples_in_cols}{\code{TRUE} if samples are in columns and features in rows, \code{FALSE} if it is reversed. See examples for more information.}
}
\value{
A tidy tibble.
}
\description{
Convert a wide matrix to a tidy tibble
}
\examples{
# Using a small fictional data set
dataset <- matrix(1:9, ncol = 3)
colnames(dataset) <- paste0("sample", 1:3)
rownames(dataset) <- paste0("feature", 1:3)


# Example 1: Samples in columns
dataset
convert_from_matrix(dataset)

# Example 2: Samples in rows
dataset_transposed <- t(dataset)
dataset_transposed

convert_from_matrix(dataset, samples_in_cols = FALSE)
}
