% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_from_wide}
\alias{convert_from_wide}
\title{Convert a wide feature table to a tidy tibble}
\usage{
convert_from_wide(data, label_col = 1, metadata_cols = NULL)
}
\arguments{
\item{data}{A feature table data frame or tibble in wide format. To convert a matrix, see \code{\link[metamorphr]{convert_from_matrix}}.}

\item{label_col}{The index or name of the column that will be used to label Features. For example an identifier (\emph{e.g.}, KEGG, CAS, HMDB) or a \emph{m/z}-RT pair.}

\item{metadata_cols}{The index/indices or name(s) of column(s) that hold additional feature metadata (\emph{e.g.}, retention times, additional identifiers or \emph{m/z} values).}
}
\value{
A tidy tibble.
}
\description{
Feature tables are usually stored in a "wide" data format where sample names are stored in columns and features are stored in rows.
This functions transforms those feature tables into a "long" and tidy data format to use it with functions provided in the \code{metamorphr} package.
\code{convert_from_wide} works with tibbles and data frames. To convert a matrix, see \code{\link[metamorphr]{convert_from_matrix}}.
}
\examples{
featuretable_path <- system.file("extdata", "toy_metaboscape.csv", package = "metamorphr")
featuretable_wide <- read.csv(featuretable_path)

convert_from_wide(featuretable_wide, metadata_cols = 2:5)

}
