% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{create_metadata_skeleton}
\alias{create_metadata_skeleton}
\title{Create a blank metadata skeleton}
\usage{
create_metadata_skeleton(data)
}
\arguments{
\item{data}{A tidy tibble created by \code{metamorph::read_featuretable()}.}
}
\value{
An empty tibble structure with the necessary columns for metadata.
}
\description{
The function takes a tidy tibble created by \code{metamorph::read_featuretable()} and returns an empty tibble for sample metadata. The tibble can either be populated directly in R or exported and edited by hand (\emph{e.g.} in Excel). Metadata are necessary for several downstream functions. More columns may be added if necessary.
}
\examples{
feature_table <- read_featuretable("path/to/sample/data.csv", metadata_cols = c(2,3,4))
metadata_table <- create_metadata_skeleton(feature_table)
}
