% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_cv}
\alias{filter_cv}
\title{Filter Features based on their coefficient of variation}
\usage{
filter_cv(
  data,
  max_cv,
  reference_samples,
  ref_as_group = FALSE,
  grouping_column = NULL,
  na_as_zero = TRUE
)
}
\arguments{
\item{data}{A tidy tibble created by \code{\link[metamorphr]{read_featuretable}}.}

\item{max_cv}{The maximum allowed CV. 0.2 is a reasonable start.}

\item{reference_samples}{The names of the samples which will be used to calculate the CV of a feature. Often Quality Control samples.}

\item{na_as_zero}{Should \code{NA} be replaced with 0 prior to calculation?
Under the hood \code{filter_cv} calculates the CV by \code{stats::sd(..., na.rm = TRUE) / mean(..., na.rm = TRUE)}.
If there are 3 samples to calculate the CV from and 2 of them are \code{NA} for a specific feature, then the CV for that Feature will be \code{NA}
if \code{na_as_zero = FALSE}. This might lead to problems. \code{na_as_zero = TRUE} is the safer pick.
Zeros will be replaced with \code{NA} after calculation no matter if it is \code{TRUE} or \code{FALSE}.}
}
\value{
A filtered tibble.
}
\description{
One of several filter functions. It filters Features based on their coefficient of variation (CV).
It is defined as \eqn{CV = \frac{\sigma }{\mu }} with \eqn{\sigma} = Standard deviation and \eqn{\mu} = Mean.
}
\examples{
toy_metaboscape \%>\%
  filter_cv(max_cv = 0.2, reference_samples = c("QC1", "QC2", "QC3"))
}
\references{
\if{html}{\out{<a href="https://en.wikipedia.org/wiki/Coefficient_of_variation">}}Coefficient of Variation on Wikipedia\if{html}{\out{</a>}}
}
