% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputes.R
\name{impute_global_lowest}
\alias{impute_global_lowest}
\title{Impute missing values by replacing them with the lowest observed intensity (global)}
\usage{
impute_global_lowest(data)
}
\arguments{
\item{data}{A tidy tibble created by \code{metamorphr::read_featuretable()}.}
}
\value{
A tibble with imputed missing values.
}
\description{
Replace missing intensity values (\code{NA}) with the lowest observed intensity. This imputation method is very rudimentary and other methods might provide better results.
}
\examples{
toy_metaboscape \%>\%
  impute_global_lowest()
}
