% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputes.R
\name{impute_min}
\alias{impute_min}
\title{Impute missing values by replacing them with the Feature minimum}
\usage{
impute_min(data)
}
\arguments{
\item{data}{A tidy tibble created by \code{\link[metamorphr]{read_featuretable}}.}
}
\value{
A tibble with imputed missing values.
}
\description{
Replace missing intensity values (\code{NA}) with the Feature minimum of non-\code{NA} values.
}
\examples{
toy_metaboscape \%>\%
  impute_min()
}
