% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputes.R
\name{impute_rf}
\alias{impute_rf}
\title{Impute missing values using random forest}
\usage{
impute_rf(data, random_seed = 1L, ...)
}
\arguments{
\item{data}{A tidy tibble created by \code{\link[metamorphr]{read_featuretable}}.}

\item{random_seed}{A seed for the random number generator. Can be an integer or \code{NULL} (in case no particular seed should be used) but for reproducibility reasons it is \strong{strongly advised} to provide an integer.}

\item{...}{}
}
\value{
A tibble with imputed missing values.
}
\description{
Basically a wrapper function around \code{missForest::missForest()}. Imputes missing values using the random forest algorithm.
}
\examples{
toy_metaboscape \%>\%
  impute_knn()
}
\references{
For further information, see\if{html}{\out{
<ul>
<li>Daniel J. Stekhoven <stekhoven@stat.math.ethz.ch>, <b>2011</b>, DOI <a href = "https://doi.org/10.32614/CRAN.package.missForest">10.32614/CRAN.package.missForest</a>.</li>
<li>D. J. Stekhoven, P. Bühlmann, <i>Bioinformatics</i> <b>2012</b>, <i>28</i>, 112–118, DOI <a href = "https://doi.org/10.1093/bioinformatics/btr597">10.1093/bioinformatics/btr597</a>.</li>
</ul>
}}
}
