% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputes.R
\name{impute_user_value}
\alias{impute_user_value}
\title{Impute missing values by replacing them with a user-provided value}
\usage{
impute_user_value(data, value)
}
\arguments{
\item{data}{A tidy tibble created by \code{\link[metamorphr]{read_featuretable}}.}

\item{value}{Numeric that replaces missing values}
}
\value{
A tibble with imputed missing values.
}
\description{
Replace missing intensity values (\code{NA}) with a user-provided value (e.g., 1).
}
\examples{
toy_metaboscape \%>\%
  impute_user_value(value = 1)
}
