% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{join_metadata}
\alias{join_metadata}
\title{Join a featuretable and sample metadata}
\usage{
join_metadata(data, metadata)
}
\arguments{
\item{data}{A feature table created with \code{\link[metamorphr]{read_featuretable}}}

\item{metadata}{Sample metadata created with \code{\link[metamorphr]{create_metadata_skeleton}}}
}
\value{
A tibble with added sample metadata.
}
\description{
Joins a featuretable and associated sample metadata. Basically a wrapper around \code{\link[dplyr]{left_join}} where \code{by = "Sample"}.
}
\examples{
toy_metaboscape \%>\%
  join_metadata(toy_metaboscape_metadata)
}
