% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamorph-package.R
\docType{package}
\name{metamorph-package}
\alias{metamorph}
\alias{metamorph-package}
\title{metamorph: Tidy and Streamlined Metabolomics Data Workflows}
\description{
This package contains functions and wrapper functions to facilitate tasks typically encountered during metabolomics data analysis including data import, filtering, missing value imputation, transformation, centering and scaling as well as some statistical tests and plotting. It offers easy integration with other Tidyverse packages including ggplot2 as data is stored in a tidy tibble.
}
\author{
\strong{Maintainer}: Yannik Schermer \email{yannik.schermer@chem.rptu.de} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
