% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msn.R
\name{msn_scale}
\alias{msn_scale}
\title{Scale intensities in MSn spectra to the highest value within each spectrum}
\usage{
msn_scale(data, scale_to = 100)
}
\arguments{
\item{data}{A tidy tibble created by \code{\link[metamorphr]{read_featuretable}}.}

\item{scale_to}{A \code{numeric} that specifies to which number the highest signal in each spectrum will be scaled.}
}
\value{
A tibble with scaled MSn spectra.
}
\description{
Scale the intensity of each peak in an MSn spectrum to that of the highest peak. MSn spectra are required to use this function.
See \code{\link[metamorphr]{read_mgf}}.

\strong{Important Note}

Please note that existing MSn spectra in \code{data} will be overwritten.
}
\examples{
toy_mgf \%>\%
  msn_scale()
}
