% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_median}
\alias{normalize_median}
\title{Normalize intensities across samples by dividing by the sample median}
\usage{
normalize_median(data)
}
\arguments{
\item{data}{A tidy tibble created by \code{\link[metamorphr]{read_featuretable}}.}
}
\value{
A tibble with intensities normalized across samples.
}
\description{
Normalize across samples by dividing feature intensities by the sample median, making the median 1 in all samples.
See References for more information.
}
\examples{
toy_metaboscape \%>\%
  normalize_median()
}
\references{
T. Ramirez, A. Strigun, A. Verlohner, H.-A. Huener, E. Peter, M. Herold, N. Bordag, W. Mellert, T. Walk, M. Spitzer, X. Jiang, S. Sperber, T. Hofmann, T. Hartung, H. Kamp, B. Van Ravenzwaay, \emph{Arch Toxicol} \strong{2018}, \emph{92}, 893–906, DOI \href{https://doi.org/10.1007/s00204-017-2079-6}{10.1007/s00204-017-2079-6}.
}
