% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_quantile_all}
\alias{normalize_quantile_all}
\title{Normalize intensities across samples using standard Quantile Normalization}
\usage{
normalize_quantile_all(data)
}
\arguments{
\item{data}{A tidy tibble created by \code{\link[metamorphr]{read_featuretable}}.}
}
\value{
A tibble with intensities normalized across samples.
}
\description{
This is the standard approach for Quantile Normalization. Other sub-flavors are also available:
\itemize{
\item \code{\link[metamorphr]{normalize_quantile_group}}
\item \code{\link[metamorphr]{normalize_quantile_batch}}
\item \code{\link[metamorphr]{normalize_quantile_smooth}}
}

See References for more information.
}
\examples{
toy_metaboscape \%>\%
  normalize_quantile_all()
}
\references{
Y. Zhao, L. Wong, W. W. B. Goh, \if{html}{\out{<i>}}Sci Rep\if{html}{\out{</i>}} \if{html}{\out{<b>}}2020\if{html}{\out{</b>}}, \if{html}{\out{<i>}}10\if{html}{\out{</i>}}, 15534, DOI \if{html}{\out{<a href = "https://doi.org/10.1038/s41598-020-72664-6">}}10.1038/s41598-020-72664-6\if{html}{\out{</a>}}.
}
