% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pca}
\alias{plot_pca}
\title{Draws a scores or loadings plot or performs calculations necessary to draw them manually}
\usage{
plot_pca(
  data,
  method = "svd",
  what = "scores",
  n_pcs = 2,
  pcs = c(1, 2),
  group_column = NULL,
  name_column = NULL,
  return_tbl = FALSE
)
}
\arguments{
\item{data}{A tidy tibble created by \code{\link[metamorphr]{read_featuretable}}.}

\item{method}{A character specifying one of the available methods ("svd", "nipals", "rnipals", "bpca", "ppca", "svdImpute", "robustPca", "nlpca", "llsImpute", "llsImputeAll"). If the default is used ("svd") an SVD PCA will be done, in case \code{data} does not contain missing values, or a NIPALS PCA if \code{data} does contain missing values.}

\item{what}{Specifies what should be returned. Either \code{"scores"} or \code{"loadings"}.}

\item{n_pcs}{The number of PCs to calculate.}

\item{pcs}{A vector containing 2 integers that specifies the PCs to plot. The following condition applies: \code{max(pcs) <= n_pcs}.}

\item{group_column}{Either \code{NULL} or a column in \code{data} (e.g., \code{group_column = Group}). If provided, the dots in the scores plot will be colored according to their group. Only relevant if \code{what = "scores"}.}

\item{name_column}{Either \code{NULL} or a column in \code{data} (e.g., \code{name_column = Feature}). If provided, feature names are preserved in the resulting tibble. Only relevant if \code{what = "loadings"} & \code{return_tbl = TRUE}.}

\item{return_tbl}{A logical. If \code{FALSE}, returns a ggplot2 object, if \code{TRUE} returns a tibble which can be used to draw the plot manually to have more control.}
}
\description{
Draws a scores or loadings plot or performs calculations necessary to draw them manually
}
