% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_featuretable}
\alias{read_featuretable}
\title{Read a feature table into a tidy tibble}
\usage{
read_featuretable(file, delim = ",", label_col = 1, metadata_cols = NULL, ...)
}
\arguments{
\item{file}{A path to a file but can also be a connection or literal data.}

\item{delim}{The field separator or delimiter. For example "," in csv files.}

\item{label_col}{The index of the column that will be used to label metabolites. For example an identifier (\emph{e.g.}, KEGG, CAS, HMDB) or a \emph{m/z}-RT pair.}

\item{metadata_cols}{The index/indices of column(s) that hold additional metadata (\emph{e.g.}, retention times, additional identifiers or \emph{m/z} values).}

\item{...}{Additional arguments passed on to \code{readr::read_delim()}}
}
\value{
A tidy tibble.
}
\description{
Basically a wrapper around \code{readr::read_delim()} but performs some initial tidying operations such as \code{gather()} rearranging columns. The \code{label_col} will be renamed to \emph{Metabolite}.
}
\examples{
featuretable_path <- system.file("extdata", "toy_metaboscape.csv", package = "metamorphr")
featuretable <- read_featuretable(featuretable_path, metadata_cols = 2:5)
}
