% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{remove_empty_cols}
\alias{remove_empty_cols}
\title{Remove empty columns from a tibble or data frame}
\usage{
remove_empty_cols(data)
}
\arguments{
\item{data}{A tibble or data frame in wide format.}
}
\value{
A tibble or data frame in wide format without empty columns.
}
\description{
Remove empty columns (i.e., columns that \emph{only} contain \code{NA}) from a tibble or data frame.
}
\examples{
# Column `d` contains only `NA` and should be removed
na_tibble <- tibble::tibble(a = c(1,2,3), b = c(NA, 2,3), c = c(NA, NA, 3), d = c(NA,NA,NA))

remove_empty_cols(na_tibble)
}
