% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_vast_grouped}
\alias{scale_vast_grouped}
\title{Scale intensities of features using grouped vast scaling}
\usage{
scale_vast_grouped(data, group_column = .data$Group)
}
\arguments{
\item{data}{A tidy tibble created by \code{\link[metamorphr]{read_featuretable}}.}

\item{group_column}{Which column should be used for grouping? Usually \code{grouping_column = Group}. Uses \code{\link[rlang]{args_data_masking}}.}
}
\value{
A tibble with vast scaled intensities.
}
\description{
Scale intensities of features using grouped vast scaling
}
\examples{
toy_metaboscape \%>\%
  join_metadata(toy_metaboscape_metadata) \%>\%
  scale_vast_grouped()
}
\references{
For further information, see\if{html}{\out{
<ul>
<li>R. A. Van Den Berg, H. C. Hoefsloot, J. A. Westerhuis, A. K. Smilde, M. J. Van Der Werf, <i>BMC Genomics</i> <b>2006</b>, <i>7</i>, 142, DOI <a href = "https://doi.org/10.1186/1471-2164-7-142">10.1186/1471-2164-7-142</a>.</li>
</ul>
}}
}
