% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_mgf}
\alias{toy_mgf}
\title{A small toy data set containing MSn spectra}
\format{
\subsection{\code{toy_mgf}}{

A data frame with 3 rows and 5 columns:
\describe{
\item{VARIABLEONE}{A fictional variable.}
\item{VARIABLETWO}{A fictional variable.}
\item{VARIABLETHREE}{A fictional variable.}
\item{PEPMASS}{The precursor ion m/z.}
\item{MSn}{A list column containing MSn spectra.}
...
}
}
}
\source{
This data set contains fictional data!
}
\usage{
toy_mgf
}
\description{
Data was generated with \code{metamorphr::read_mgf()} and can be reproduced with
This tibble can be reproduced with \code{metamorphr::read_mgf(system.file("extdata", "toy_mgf.mgf", package = "metamorphr"))}.
}
\keyword{datasets}
