% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputes.R
\name{impute_knn}
\alias{impute_knn}
\title{Impute missing values using nearest neighbor averaging}
\usage{
impute_knn(data, quietly = TRUE, ...)
}
\arguments{
\item{data}{A tidy tibble created by \code{\link[metamorphr]{read_featuretable}}.}

\item{quietly}{\code{TRUE} or \code{FALSE}. Should messages and warnings from \code{\link[impute]{impute.knn}} be printed to the console?}

\item{...}{Additional parameters passed to \code{\link[impute]{impute.knn}}.}
}
\value{
A tibble with imputed missing values.
}
\description{
Basically a wrapper function around \verb{impute::}\code{\link[impute]{impute.knn}}. Imputes missing values using the k-th nearest neighbor algorithm.

Note that the function ln-transforms the data prior to imputation and transforms it back to the original scale afterwards. \strong{Please do not do it manually prior to calling \code{impute_knn()}!}
See References for more information.

\if{html}{\out{<b>}}Important Note\if{html}{\out{</b>}}\if{html}{\out{<br>}}
\code{impute_knn()} depends on the \code{impute} package from Bioconductor. If \code{metamorphr} was installed via \code{install.packages()}, dependencies from Bioconductor were not
automatically installed. When \code{impute_knn()} is called without the \code{impute} package installed, you should be asked if you want to install \code{pak} and \code{impute}.
If you want to use \code{impute_knn()} you have to install those. In case you run into trouble with the automatic installation, please install \code{impute} manually. See
\if{html}{\out{<a href = "https://bioconductor.org/packages/release/bioc/html/impute.html">}}impute: Imputation for microarray data\if{html}{\out{</a>}} for instructions on manual installation.
}
\examples{
toy_metaboscape \%>\%
  impute_knn()
}
\references{
For further information, see\if{html}{\out{
<ul>
<li>Robert Tibshirani, Trevor Hastie, <b>2017</b>, DOI <a href = "https://doi.org/10.18129/B9.BIOC.IMPUTE">10.18129/B9.BIOC.IMPUTE</a>.</li>
<li>J. Khan, J. S. Wei, M. Ringnér, L. H. Saal, M. Ladanyi, F. Westermann, F. Berthold, M. Schwab, C. R. Antonescu, C. Peterson, P. S. Meltzer, <i>Nat Med</i> <b>2001</b>, <i>7</i>, 673–679, DOI <a href = "https://doi.org/10.1038/89044">10.1038/89044</a>.</li>
</ul>
}}
}
