% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamorphr-package.R
\docType{package}
\name{metamorphr-package}
\alias{metamorphr}
\alias{metamorphr-package}
\title{metamorphr: Tidy and Streamlined Metabolomics Data Workflows}
\description{
Facilitate tasks typically encountered during metabolomics data analysis including data import, filtering, missing value imputation, transformation, centering and scaling as well as statistical tests and plotting. 'metamorphr' introduces a tidy format for metabolomics data and is designed to make it easier to build elaborate analysis workflows and to integrate them with 'Tidyverse' packages including 'dplyr' and 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yasche/metamorphr}
  \item Report bugs at \url{https://github.com/yasche/metamorphr/issues}
}

}
\author{
\strong{Maintainer}: Yannik Schermer \email{yannik.schermer@chem.rptu.de} (\href{https://orcid.org/0009-0002-5201-057X}{ORCID}) [copyright holder]

}
\keyword{internal}
