% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_featuretable_mzmine}
\alias{read_featuretable_mzmine}
\title{Read a 'full_feature_table' from 'mzmine' into a tidy tibble}
\usage{
read_featuretable_mzmine(
  file,
  intensity = "height",
  field_separator = ",",
  label_col = 1,
  import_datafile_cols = FALSE,
  remove_empty_cols = FALSE,
  show_removed_cols = TRUE
)
}
\arguments{
\item{file}{A path to a file but can also be a connection or literal data.}

\item{intensity}{A character that specifies what should be used as the (semi-)quantitative measure. Either \code{"height"} or \code{"area"}.}

\item{field_separator}{The field separator as specified in 'mzmine'. Usually \code{","} if the file is in common CSV format.}

\item{label_col}{The index or name of the column that will be used to label Features. For example an identifier (\emph{e.g.}, KEGG, CAS, HMDB) or a \emph{m/z}-RT pair.}

\item{import_datafile_cols}{Should columns that begin with \verb{datafile:} be imported? Those columns contain sample-specific information, for example the retention time of a feature measured in a specific sample. Usually, this information is not necessary for downstream analysis but it can be used for quaility control purposes. If \code{TRUE}, \verb{datafile:} columns are imported and the sample names are removed from the column names. This allows for tidy storage of the information in one column per variable.}

\item{remove_empty_cols}{Either \code{TRUE} or \code{FALSE}. Should empty columns be removed after reading the feature table? For a more fine-grained control, you can use a combination of \code{\link[readr]{read_delim}}, \code{\link[metamorphr]{remove_empty_cols}} and \code{\link[metamorphr]{convert_from_wide}}.See the respective function documentation for more details.}

\item{show_removed_cols}{Only relevant if \code{remove_empty_cols = TRUE}. If \code{TRUE} prints a message that shows which columns were removed.}
}
\value{
A tidy tibble.
}
\description{
Similar to \code{\link[metamorphr]{read_featuretable}} but specifically for full_feature_table' files created with 'mzmine'. For more information, see \href{https://mzmine.github.io/mzmine_documentation/module_docs/io/feat-list-export.html}{the 'mzmine' documentation}.
}
\examples{
# Read a toy dataset in the format produced with mzmine.

featuretable_path <- system.file("extdata", "toy_mzmine.csv", package = "metamorphr")

# Example 1: Use feature height as the metric
featuretable <- read_featuretable_mzmine(
  featuretable_path,
  intensity = "height"
)

featuretable

# Example 2: Use feature area as the metric
featuretable <- read_featuretable_mzmine(
  featuretable_path,
  intensity = "area"
)

featuretable

# Example 3: Use the 'mz' column as a Feature label
featuretable <- read_featuretable_mzmine(
  featuretable_path,
  label_col = "mz"
)

featuretable
}
\references{
\itemize{
\item H. Wickham, \emph{J. Stat. Soft.} \strong{2014}, \emph{59}, DOI 10.18637/jss.v059.i10.
\item H. Wickham, M. Averick, J. Bryan, W. Chang, L. McGowan, R. François, G. Grolemund, A. Hayes, L. Henry, J. Hester, M. Kuhn, T. Pedersen, E. Miller, S. Bache, K. Müller, J. Ooms, D. Robinson, D. Seidel, V. Spinu, K. Takahashi, D. Vaughan, C. Wilke, K. Woo, H. Yutani, \emph{JOSS} \strong{2019}, \emph{4}, 1686, DOI 10.21105/joss.01686.
\item “12 Tidy data | R for Data Science,” can be found under \url{https://r4ds.had.co.nz/tidy-data.html}, \strong{2023}.
}
}
