% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_log}
\alias{transform_log}
\title{Transforms the intensities by calculating their log}
\usage{
transform_log(data, base = 10)
}
\arguments{
\item{data}{A tidy tibble created by \code{\link[metamorphr]{read_featuretable}}.}

\item{base}{Which base should be used for the log-transformation. The default (10) means that log\if{html}{\out{<sub>}}10\if{html}{\out{</sub>}} values of the intensities are calculated.}
}
\value{
A tibble with log-transformed intensities.
}
\description{
Transforms the intensities by calculating their log
}
\examples{
toy_metaboscape \%>\%
  impute_lod() \%>\%
  transform_log()
}
